requirements.txt
beautifulsoup4==4.13.3
blinker==1.8.2
bs4==0.0.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.5
pdf2image==1.17.0
pika==1.3.2
pillow==11.0.0
psutil==6.1.0
reportlab==4.2.5
soupsieve==2.6
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.2
